+++ R/targets_misc.R
@@ -19,8 +19,9 @@
 #' @export
 #'
 #' @examples
-#' list_fastq_files("extdata")
-#' list_fastq_files("extdata", paired_end = FALSE, pattern_R1 = "")
+#' list_fastq_files(system.file("extdata", package = "MiscMetabar"))
+#' list_fastq_files(system.file("extdata", package = "MiscMetabar"),
+#'                               paired_end = FALSE, pattern_R1 = "")
 #'
 #' @author Adrien Taudière
 
@@ -173,36 +174,25 @@ filter_trim <-
         )
         dir.create(output_rev)
         dir.create(output_fw)
-        file.rename(
-          paste0(output_rev, "interm"),
-          paste0(output_rev, "/", basename(rev))
-        )
-        file.rename(
-          paste0(output_fw, "interm"),
-          paste0(output_fw, "/", basename(fw))
-        )
+        file.rename(paste0(output_rev, "interm"),
+                    paste0(output_rev, "/", basename(rev)))
+        file.rename(paste0(output_fw, "interm"),
+                    paste0(output_fw, "/", basename(fw)))
 
         return(list("fw" = output_fw, "rv" = output_rev))
       } else {
-        dada2::filterAndTrim(
-          filt = paste0(output_fw, "interm"),
-          fwd = fw,
-          ...
-        )
+        dada2::filterAndTrim(filt = paste0(output_fw, "interm"),
+                             fwd = fw,
+                             ...)
         dir.create(output_fw)
 
-        file.rename(
-          paste0(output_fw, "interm"),
-          paste0(output_fw, "/", basename(fw))
-        )
+        file.rename(paste0(output_fw, "interm"),
+                    paste0(output_fw, "/", basename(fw)))
         return(output_fw)
       }
     } else {
       if (is.null(rev)) {
-        dada2::filterAndTrim(
-          filt = output_fw,
-          fwd = fw, ...
-        )
+        dada2::filterAndTrim(filt = output_fw, fwd = fw, ...)
         return(output_fw)
       } else {
         dada2::filterAndTrim(
@@ -300,7 +290,8 @@ sample_data_with_new_names <- function(file_path,
 rename_samples <- function(phyloseq_component,
                            names_of_samples,
                            taxa_are_rows = FALSE) {
-  if (is.null(sample_names(phyloseq_component)) && inherits(phyloseq_component, "matrix")) {
+  if (is.null(sample_names(phyloseq_component)) &&
+      inherits(phyloseq_component, "matrix")) {
     phyloseq_component <- otu_table(phyloseq_component, taxa_are_rows = taxa_are_rows)
   }
   if (length(names_of_samples) != length(sample_names(phyloseq_component))) {
@@ -333,33 +324,31 @@ rename_samples <- function(phyloseq_component,
 #'   fastq files names.
 #' @param verbose (logical, default TRUE) If TRUE, print some additional messages.
 #' @param remove_undocumented_fastq_files (logical, default FALSE) If set to TRUE
-#'   fastq files not present in sam_data are removed from your folder. 
-#'   Keep a copy of those files somewhere before. 
-#' @param prefix Add a prefix to new samples names (ex. prefix = "samp") 
+#'   fastq files not present in sam_data are removed from your folder.
+#'   Keep a copy of those files somewhere before.
+#' @param prefix Add a prefix to new samples names (ex. prefix = "samp")
 #' @param ... Other parameters passed on to [utils::read.csv()] function.
 #' @return A list of two objects :
 #'   - $sam_names_matching is a tibble of corresponding samples names
-#'   - $sam_data is a sample data files including only matching sample names 
+#'   - $sam_data is a sample data files including only matching sample names
 #' @importFrom utils read.csv
 #' @export
 #' @author Adrien Taudière
-sam_data_matching_names <- function(
-  path_sam_data, 
-  sample_col_name, 
-  path_raw_seq, 
-  pattern_remove_sam_data = NULL,
-  pattern_remove_fastq_files = NULL,
-  verbose = TRUE,
-  remove_undocumented_fastq_files = FALSE,
-  prefix = NULL,
-  ...
-) {
+sam_data_matching_names <- function(path_sam_data,
+                                    sample_col_name,
+                                    path_raw_seq,
+                                    pattern_remove_sam_data = NULL,
+                                    pattern_remove_fastq_files = NULL,
+                                    verbose = TRUE,
+                                    remove_undocumented_fastq_files = FALSE,
+                                    prefix = NULL,
+                                    ...) {
   sam_d <- read.csv(path_sam_data, ...)
   names_sam_data <- sam_d[[sample_col_name]]
   names_fastq_files_fullpath <- list.files(path_raw_seq)
   names_fastq_files <- basename(names_fastq_files_fullpath)
 
-  if(!is.null(pattern_remove_sam_data)) {
+  if (!is.null(pattern_remove_sam_data)) {
     names_sam_data_clean <- gsub(pattern_remove_sam_data, "", names_sam_data)
     if (sum(duplicated(names_sam_data_clean)) > 0) {
       stop(
@@ -367,19 +356,23 @@ sam_data_matching_names <- function(
         names_sam_data_clean[duplicated(names_sam_data_clean)]
       )
     }
+  } else {
+    names_sam_data_clean <- names_sam_data
   }
 
-  if(!is.null(pattern_remove_fastq_files)) {
+  if (!is.null(pattern_remove_fastq_files)) {
     names_fastq_files_clean <- gsub(pattern_remove_fastq_files, "", names_fastq_files)
+  } else {
+    names_fastq_files_clean <- names_fastq_files
   }
 
   tib_sam_data <- tibble(
-    clean_sam = as.character(names_sam_data_clean), 
+    clean_sam = as.character(names_sam_data_clean),
     raw_sam = as.character(names_sam_data)
   )
 
   tib_fastq <- tibble(
-    clean_fastq = as.character(names_fastq_files_clean), 
+    clean_fastq = as.character(names_fastq_files_clean),
     raw_fastq = as.character(names_fastq_files),
     raw_fastq_full_path = names_fastq_files_fullpath
   )
@@ -387,35 +380,50 @@ sam_data_matching_names <- function(
   tib_j <- full_join(tib_fastq, tib_sam_data, by = join_by(clean_fastq == clean_sam)) |>
     rename(common_names = clean_fastq)
 
-  if(sum(is.na(tib_j$raw_fastq))>0){
-    message(sum(is.na(tib_j$raw_fastq)), " samples in sam_data files are not present in fastq_files")
-    if(verbose) {
-      warning(tib_j$raw_sam[is.na(tib_j$raw_fastq)], "not_matching_names_from_sam_data.txt")
+  if (sum(is.na(tib_j$raw_fastq)) > 0) {
+    message(sum(is.na(tib_j$raw_fastq)),
+            " samples in sam_data files are not present in fastq_files")
+    if (verbose) {
+      warning(tib_j$raw_sam[is.na(tib_j$raw_fastq)],
+              "not_matching_names_from_sam_data.txt")
     }
   }
 
-  if(sum(is.na(tib_j$raw_sam))>0){
-    message(sum(is.na(tib_j$raw_sam)), " samples in fastq files are not present in sam_data")
-    if(verbose) {
-      warning(tib_j$raw_fastq[is.na(tib_j$raw_sam)], "not_matching_names_from_fastq_files.txt")
+  if (sum(is.na(tib_j$raw_sam)) > 0) {
+    message(sum(is.na(tib_j$raw_sam)),
+            " samples in fastq files are not present in sam_data")
+    if (verbose) {
+      warning(tib_j$raw_fastq[is.na(tib_j$raw_sam)],
+              "not_matching_names_from_fastq_files.txt")
     }
-    if(remove_undocumented_fastq_files){
-      if(verbose) {
-        warning("Files ", tib_j$raw_fastq_full_path[is.na(tib_j$raw_sam)], "will be removed from folder", path_raw_seq)
+    if (remove_undocumented_fastq_files) {
+      if (verbose) {
+        warning(
+          "Files ",
+          tib_j$raw_fastq_full_path[is.na(tib_j$raw_sam)],
+          "will be removed from folder",
+          path_raw_seq
+        )
       }
       unlink(tib_j$raw_fastq_full_path[is.na(tib_j$raw_sam)])
     }
   }
-  
-  if(!is.null(prefix)){
+
+  if (!is.null(prefix)) {
     tib_j$common_names <- paste0(prefix, tib_j$common_names)
   }
 
-  sam_d_new <- sam_d |>
-    dplyr::filter(.data[[sample_col_name]] %in% tib_j$raw_sam[!is.na(tib_j$raw_fastq)]) |>
-    dplyr::mutate("samples_names_common" = paste0(prefix, gsub(pattern_remove_sam_data, "", .data[[sample_col_name]]))) |>
+  if (is.null(pattern_remove_sam_data)) {
+    sam_d_new <- sam_d |>
+      dplyr::filter(.data[[sample_col_name]] %in% tib_j$raw_sam[!is.na(tib_j$raw_fastq)]) |>
+      dplyr::mutate("samples_names_common" = paste0(prefix, .data[[sample_col_name]])) |>
       relocate(samples_names_common)
-
+  } else {
+    sam_d_new <- sam_d |>
+      dplyr::filter(.data[[sample_col_name]] %in% tib_j$raw_sam[!is.na(tib_j$raw_fastq)]) |>
+      dplyr::mutate("samples_names_common" = paste0(prefix, gsub(pattern_remove_sam_data, "", .data[[sample_col_name]]))) |>
+      relocate(samples_names_common)
+  }
   return(list("sam_names_matching" = tib_j, "sam_data" = sam_d_new))
 }
 ################################################################################