+++ NEWS.md
@@ -1,7 +1,8 @@
 # MiscMetabar 0.10.3 (in development)
 
-- Add params `type`, `na_remove` and `verbose` to `ggvenn_pq()`. The type = "nb_seq" allow to plot Venn diagramm with the number of shared sequences instead of shared ASV. 
+- Add params `type`, `na_remove` and `verbose` to `ggvenn_pq()`. The type = "nb_seq" allow to plot Venn diagram with the number of shared sequences instead of shared ASV. 
 - Add automatic report in json for the function `cutadapt_remove_primers()`.
+- Add param `verbose` to `track_wkflow()` and improve examples for `track_wkflow()` and `list_fastq_files`
 
 # MiscMetabar 0.10.2 (in development)
 
@@ -20,7 +21,7 @@
 # MiscMetabar 0.9.4 
 
 - Set a seed in the example of `build_tree_pq` to resubmit to CRAN
-   Add a param `return_a_vector` in function `filter_trim()` to make possible to return a vector of path as it is usefull when used with `targets::tar_targets(..., format="file")`)
+   Add a param `return_a_vector` in function `filter_trim()` to make possible to return a vector of path as it is useful when used with `targets::tar_targets(..., format="file")`)
 - Make some storage amelioration by replacing `list()` by `vector(list, ...)` 
 
 # MiscMetabar 0.9.3